pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyegm"
version = "0.2.5"
authors = [{name = "CJiangqiu", email = "17787153839@qq.com"}]
description = "Explosive Generative Model (EGM) classifier"
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "typing-extensions>=4.0.0; python_version < '3.8'",
    "scipy>=1.5.0",
    "hnswlib>=0.5.2",
    "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/CJiangqiu/PyEGM"
Documentation = "https://github.com/CJiangqiu/PyEGM#readme"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pyegm = ["data/*.yaml", "data/*.yml", "data/*.json"]
